Organize base discover page codes.
