+++ together-wherever/src/constants/defaultCenter.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_CENTER = {
+    lat: 13.736717, // Default to Bangkok, Thailand
+    lng: 100.523186,
+};
